rmq-kafka-worker/requirements.txt
kombu==5.2.*
kafka-python
avro-python3